+++ inc/ticket.class.php
@@ -102,6 +102,24 @@ class PluginEscaladeTicket
                 }
             }
         }
+        if (!isset($item->input['actortype'])) {
+            $groups = new Group_Ticket();
+            $groups = $groups->find(['tickets_id' => $item->getID(), 'type' => CommonITILActor::ASSIGN]);
+            foreach ($item->input['_actors']['assign'] as $actors) {
+                if ($actors['itemtype'] == 'Group' && !in_array($actors['items_id'], $groups)) {
+                    $item->input['groups_id'] = $actors['items_id'];
+                    $item->input['actortype'] = CommonITILActor::ASSIGN;
+                }
+            }
+        }
+        if (
+            (isset($item->input['actortype']) && $item->input['actortype'] == CommonITILActor::ASSIGN)
+        ) {
+            //disable notification to prevent notification for old AND new group
+            $item->input['_disablenotif'] = true;
+            return PluginEscaladeTicket::addHistoryOnAddGroup($item);
+        }
+        return $item;
     }
 
     /**
@@ -299,24 +317,12 @@ class PluginEscaladeTicket
         }
 
         //if group sent is not an assign group, return
-        if ($item->input['type'] != CommonITILActor::ASSIGN) {
+        if ($item->input['actortype'] != CommonITILActor::ASSIGN) {
             return;
         }
 
-        $tickets_id = $item->input['tickets_id'];
+        $tickets_id = $item->input['id'];
         $groups_id  = $item->input['groups_id'];
-
-        //if group already assigned, return
-        $group_ticket = new Group_Ticket();
-        $condition = [
-            'tickets_id' => $tickets_id,
-            'groups_id'  => $groups_id,
-            'type'       => CommonITILActor::ASSIGN
-        ];
-        if ($group_ticket->find($condition)) {
-            unset($_SESSION['plugin_escalade']['keep_users']);
-            return;
-        }
 
         $item->fields['status'] = CommonITILObject::ASSIGNED;
 
@@ -498,23 +504,24 @@ class PluginEscaladeTicket
             'type'       => CommonITILActor::ASSIGN
         ];
         if (!$group_ticket->find($condition)) {
-            // add group to ticket
-            $group_ticket_input = [
-                'type'       => CommonITILActor::ASSIGN,
-                'groups_id'  => $groups_id,
-                'tickets_id' => $tickets_id,
-                '_plugin_escalade_no_history' => true, // Prevent a duplicated task to be added
+            $ticket = new Ticket();
+            $ticket->getFromDB($tickets_id);
+
+            // Update the ticket with actor data in order to execute the necessary rules
+            $_form_object = [
+                '_do_not_compute_status' => true,
             ];
-
-            //handle status behavior
             if ($_SESSION['plugins']['escalade']['config']['ticket_last_status'] != -1) {
-                $group_ticket_input['_from_object']['status'] = $_SESSION['plugins']['escalade']['config']['ticket_last_status'];
+                $_form_object['status'] = $_SESSION['plugins']['escalade']['config']['ticket_last_status'];
             }
-
-            $group_ticket_input['_from_object']['_do_not_compute_status'] = true;
-
-            $group_ticket = new Group_Ticket();
-            $group_ticket->add($group_ticket_input);
+            $updates_ticket = new Ticket();
+            $updates_ticket->update($_POST['ticket_details'] + [
+                '_actors' => PluginEscaladeTicket::getTicketFieldsWithActors($tickets_id, $groups_id),
+                '_plugin_escalade_no_history' => true, // Prevent a duplicated task to be added
+                'actortype' => CommonITILActor::ASSIGN,
+                'groups_id' => $groups_id,
+                '_form_object' => $_form_object,
+            ]);
         }
 
         if (!$full_history) {
@@ -1016,6 +1023,69 @@ class PluginEscaladeTicket
         }
 
         return $itemtypes;
+    }
+
+    /**
+     * Get ticket field with actors inputs
+     *
+     * @param int $tickets_id
+     * @param int $group_id
+     *
+     * @return array
+     */
+    public static function getTicketFieldsWithActors($tickets_id, $group_id)
+    {
+        $ticket = new Ticket();
+        $link_class = [
+            'User' => new $ticket->userlinkclass(),
+            'Group' => new $ticket->grouplinkclass(),
+            'Supplier' => new $ticket->supplierlinkclass(),
+        ];
+        $ticket_actors = [];
+        $actor_types = [
+            CommonITILActor::ASSIGN => 'assign',
+            CommonITILActor::OBSERVER => 'observer',
+            CommonITILActor::REQUESTER => 'requester',
+        ];
+
+        foreach ($link_class as $itemtype => $class) {
+            $ticket_actor = $class->getActors($tickets_id);
+            foreach ($ticket_actor as $type => $value) {
+                $actors_input = [];
+                foreach ($value as $val) {
+                    $actors_input[] = [
+                        'itemtype' => $itemtype,
+                        'items_id' => $val[strtolower($itemtype . 's_id')],
+                    ];
+                }
+                $actortype = $actor_types[$type] ?? '';
+                if ($actortype) {
+                    $ticket_actors[$itemtype][$actortype] = $actors_input;
+                }
+            }
+        }
+        $ticket_actors['Group']['assign'][] = [
+            'itemtype' => 'Group',
+            'items_id' => $group_id,
+        ];
+
+        $_actors['assign'] = array_merge(
+            $ticket_actors['User']['assign'] ?? [],
+            $ticket_actors['Group']['assign'] ?? [],
+            $ticket_actors['Supplier']['assign'] ?? []
+        );
+        $_actors['observer'] = array_merge(
+            $ticket_actors['User']['observer'] ?? [],
+            $ticket_actors['Group']['observer'] ?? [],
+            $ticket_actors['Supplier']['observer'] ?? []
+        );
+        $_actors['requester'] = array_merge(
+            $ticket_actors['User']['requester'] ?? [],
+            $ticket_actors['Group']['requester'] ?? [],
+            $ticket_actors['Supplier']['requester'] ?? []
+        );
+
+        return $_actors;
     }
 
     public function showForm($ID, $options = [])