+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unrelease] -
+
+### Added
+
+- Add config option for default assignation
+
+### Fixed
+
+- Fix rules execution before escalation
+- Set assign as observer unchecked by default
+- Fixed ```blocking of user deletion```
+
+
 ## [2.9.6] - 2024-05-17
 
 ### Fixed