+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix rules execution before escalation
 - Set assign as observer unchecked by default
 - Fixed ```blocking of user deletion```
-
+- Fix ```use technician group``` option
 
 ## [2.9.6] - 2024-05-17